chore(esbuild): rm inline graceful-fs in dev-server
update the building of the dev-server such that we no longer inline
graceful-fs in the dev-server. this saves us approximately 10kb in
total - 20kb get removed from dev-server/server-process.js, and 10kb are
added to `sys/node/index.ts`.

this is accomplished by removing graceful-fs from the shared alias list,
as it's only needed in `sys` and `dev-server`. we them mark the library
as "external" so that esbuild resolves to the version held on to by the
`sys` submodule