chore(build): port d.ts bundling for mock-doc from rollup to esbuild
In #5012 when we put together a script to build `mock-doc/` with Esbuild
we refactored a custom, hand-built `.d.ts` bundler that was used in the
Rollup build to instead use `dts-bundle-generator`, a library that we
depend on elsewhere. Although in theory a better option for avoiding NIH
syndrome the output is just too different, so we want to switch to using
the old rollup approach in our new, Esbuild-based build script for
`mock-doc/`.

To accomplish this we can just export the existing function from the
Rollup script and then call it in the Esbuild script.

STENCIL-1246